Dirty fix for error in rendering page in subpath
When URL is called directrly and the app is running in subpath
(i.e. /gw/travelplanner/) the page was not rendered correctly due to an
issue in setting baseurl in route.js for this version of react.
The quick and dirty fix is: for all unregognized path, go to home and
try to parse URL parameters. It seems to work fine in most cases.
At worst, go to home and doesn't display ugly pages.